requirements.txt
streamlit
pandas
PyGithub
requests